Move data path to main.py
Requiring a path to be passed to the data loader. This allows us to keep everything that may change across environments in one file, minimizing the cognitive load for someone getting started. We should eventually replace all hard-coded parameters that we'd like to be able to change with command line args.